Add a leader() method to return the canonical VarInfo from this.
Change the format_simplify method so that it is really saying what we
want it to.  Specifically, when A and B are pointers, don't say
(EQ A B); instead say (EQ (hash A) (hash B)).  If we said the former,
Simplify would presume that A and B were always interchangeable, which
is not the case when your programming language involves mutation.  The
is_reference() check should probably be moved into ProglangType, but I
want to make sure it is correctly implemented first.