Removed the "compare_byte_char_to_integer" flag, and made byte and char always comparable to int.
I added this flag a month ago, and set it to false, because I didn't
want to see invariants comparing integers to characters.  Comparing 15
to 'm' didn't seem to make much sense.  I thought this change would
only affect the comparability.

However, it is assumed many places in the code that there are only two
types of integers: integers and pointers.  For instance, !isInteger()
is used to check if a variable is a pointer.

By making byte and char not comparable to int, I essentially added a
third type of integer: characters.  This caused problems, because
characters were being confused with pointers.

I have reverted Daikon to its previous state, where byte and char are
comparable to int.  The problem of chars being compared to ints will
go away once we start using Lackwit for our C programs.