These are a bunch of utility files for randomly selecting k elements from an Iterator of size n using O(k) space, as opposed to the more intuitive O(n) space.  You also do not know the value of 'n'.
MultiRandSelector takes in a procedural abstraction, EquivalenceChecker,
which is an interface with one method, "deriveEquivalence" which takes
in an Object of type A and outputs an Object of type B.

MultiRandSelector now chooses conceptually splits all of the n elements
into X bins, where X is the number of unique .equals() Objects of type
B ever spit out by "deriveEquivalence".  Now, MultiRandSelector will
chooses k elements from each of the X bins, totalling in k*X total
chosen elements from n.

Why do we care?  Because Dtrace files have many many invocations, often
more than enough to run the JVM out of memory even at high memory setting
levels.  Also, we do not know beforehand the "n" in Dtrace files.

As for the MultiRandSelector, the Dtrace files are filled with invocations, which are the Objects of type A.  DtraceEquivalenceChecker spits out the program point name of that invocation.  That way, there will be k invocation chosen for each program point.  That's exactly what the random selection technique needed in the last paper, and we ran into many problems with memory because we used to
load the entire dtrace into a big List.  Oh, and also only ENTER program points are chosen for random selection.  The class contains a method for finding the EXIT invocations given the ENTER invocation, and patches up the StringBuffers accordingly.

For testing, DtraceECTester seems to work very well on existing dtrace files
I've been using.  Unfortunately, depending on nonce's for the patching
uncovered a previous bug.  Simply doing a "cat" to combine dtrace files might
not be the best idea, since the nonce counter resets to 0 with each dtrace
and the resulting dtrace is technically invalid, since many invocations will
share the same nonce.  Maybe I'll write a "CatDtrace" utility that
properly increments the nonces when combining dtraces.  Does that already
exist?