Invariant.java: Added isGuardingPredicate variable. Renamed due to Implication change. Added code for invariant guarding.
Implication.java:
Refactored to make this extend the new Joiner class. Renamed variables
and removed functions (their functionality was moved into joiner)
accordingly.

Joiner.java:
New Invariant made to represent the joining of any two invariants in some way.
(eg. A, B are invariants: A and B or A -> B).