Guard for identical VarInfos flowing to each other, a problem found with static fields.  For example, there's a class B with non-static field a and static field b.  a and b are type B.  The decls file for B would look like:
B:::OBJECT
this
this.a
b
b.a

when we (correctly) relate this.a to this, the transforms are:

B:::OBJECT
this=>this.a
this.a=>this.a.a
b=>b
b.a=>b.a

Now, since b is an orphan, it relates to itself, which is wrong.  Thus
we guard for this condition.

Comment/debug changes.