Changed command line parsing methods to use GetOpt:Long, which allows easier parsing of command line options. (This module comes bundled with Perl 5). Much easier than using my own loops.
Changed convertcsv.pl to accept a
.decls file as input. In this way,
the declared type of a variable
can be set by the user.  User-provided
.decls file does not need to have
its variables in the same order as
the .csv file. (test suite for this
coming up soon.)

Added some additional checks and messages
to handle cases when there are more variable values
than variable names or vice versa.

Abstracted out subroutines for parsing
and formatting variable names.