Changes to the Java-family formats.
I switched to having the emitted Java code use reflection to access
fields of objects. This makes it possible to follow non-public fields.

I modified the design of the Quant helper methods so that they do not
throw execptions when given illegal arguments. The reason is that I
want the guarantee, to the greatest extent possible, that
invariant-checking code does not throw exceptions. This simplifies the
model that a user of tools that use Java-formatted invariants must
keep in mind when inserting invariants as code.