Fixed bug which caused a duplicate decls declaration when instrumenting javac
In general, the same class cannot have 2 methods differing only by return type:
public int x();
public float y();
-------> compile error

Likewise, if you override a method, it cannot have a different return type:
public class A
{
public float x();
}

public class B extends A
{
public int x(); -----> compile error
}


HOWEVER, you can override a method which returns a subtype of the original.  Thus:

public class A
{
public String x();
}

public class B
{
public CapString x();
}

is legal where CapString is a subclass of String



The complication is that it seems that the JVM uses the FULL method signature to invoke methods.
Thus, what happens if we do:
A myB = new B();
String myString = myB.x();

Which version of x() will this call?  We know that it should and will call B.x(), but how does the JVM know to do this if it uses the full signature.  That is, it is looking for some method with which takes no parameters, and returns a String, not a CapString.

To circumvent this issue, the compiler actually inserts an additional method in B with signature: public String x().  This new method simply calls B's original method, public CapString x().
Thus, while it is a compiler error to have 2 methods with the same name and parameters but different return types in the same class in a jave source file, it is completely legal in a java class file!  We assumed this was not the case, and hence a bug.