initial submission of a class for combining basic blocks.
The input is a graph structure (read as an adjacency list); its nodes
are basic blocks and its edges represent successor relationships.

The main method which describes the steps of the algorithm is
`mergeBasicBlocks'. The rest of the methods are just helper methods.



the information about the successor relationships between