[Chicory] Exception Handling version1
The First Version of the EH is running.
A Throw-Statement is now seen as an subexit.
The PPT will be written. and the Dtrace will be Written.

[Chicory] Exception Handling version 2

If the exceptionHandling-Parameter is set,
every Method is encapsulated by an TryCatch-Instruction.
Is the exceptionHandling-Parameter not set, only a Throw-Statement
writes an Throw in the DTrace.
Is the exceptionHandling-Parameter set, every Throw-Statement is only
recognized as an Throws in DTrace

[Chicory] Exception Handling version 3

If the exceptionHandling-Parameter is set,
every Method is encapsulated by an TryCatch-Instruction.

Chicory now differentiate between thrown and forwarded Exceptions.
Every throw-command results in a distinct PPT.

[Chicory] Exception Handling version 4

Chicory calculates now a new Stackmap table entry for the EH-TryCatch
jump instead adding a predefined Entry

+ Changed the Version Number of Daikon

[xceptonHandle] make some cleaning and Forking

[Chicory] Update to BCEL6

Update Exception-Handling-support to use BCEL After Rebase to newer
Version

removed wrong comment

[Chicory] Exception Handling version 5

changed the type of the internal variable of the exception from
java/lang/Exception to java/lang/Throwable.
After failing the DataStructures.CursorList-Test.

The synchronized-support is now compiling again.
working with Exception-Handling???
synchronized is throwing an Exception in bytecode but ??

[Chicory] Exception Handling version 6

Added filter functionality to throw(.*)-PPTs

rebase to master chicory

Fix some merging bugs